New PriceFetcher component
