+++ app/translations/es.json
@@ -257,7 +257,7 @@
     "pools": "grupos...",
     "transactions": "Transacciones",
     "sure": "¿Seguro? ",
-    "minago": " min. atrás"
+    "minago": " min."
   },
   "report": {
     "title": "Informe de Sincronización",
