+++ components/Info/Info.tsx
@@ -1,9 +1,8 @@
 /* eslint-disable react-native/no-inline-styles */
-import React, { useContext, useEffect, useState } from 'react';
-import { View, ScrollView, SafeAreaView, Image, TouchableOpacity } from 'react-native';
+import React, { useContext } from 'react';
+import { View, ScrollView, SafeAreaView, Image } from 'react-native';
 import { useTheme } from '@react-navigation/native';
 
-import RPC from '../../app/rpc';
 import RegText from '../Components/RegText';
 import ZecAmount from '../Components/ZecAmount';
 import Button from '../Button';
@@ -11,10 +10,7 @@ import Utils from '../../app/utils';
 import DetailLine from './components/DetailLine';
 import { ThemeType } from '../../app/types';
 import { ContextLoaded } from '../../app/context';
-import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome';
-import { faRefresh } from '@fortawesome/free-solid-svg-icons';
-import moment from 'moment';
-import FadeText from '../Components/FadeText';
+import PriceFetcher from '../Components/PriceFetcher';
 
 type InfoProps = {
   closeModal: () => void;
@@ -25,23 +21,6 @@ const Info: React.FunctionComponent<InfoProps> = ({ closeModal, setZecPrice }) =
   const context = useContext(ContextLoaded);
   const { info, totalBalance, translate, currency, zecPrice } = context;
   const { colors } = useTheme() as unknown as ThemeType;
-  const [refreshSure, setRefreshSure] = useState(false);
-  const [refreshMinutes, setRefreshMinutes] = useState(0);
-
-  useEffect(() => {
-    const fn = () => {
-      if (zecPrice.date > 0) {
-        const date1 = moment();
-        const date2 = moment(zecPrice.date);
-        setRefreshMinutes(date1.diff(date2, 'minutes'));
-      }
-    };
-
-    fn();
-    const inter = setInterval(fn, 5000);
-
-    return () => clearInterval(inter);
-  }, [zecPrice.date]);
 
   return (
     <SafeAreaView
@@ -124,59 +103,7 @@ const Info: React.FunctionComponent<InfoProps> = ({ closeModal, setZecPrice }) =
                     : `$ -- ${currency} per ${info.currencyName ? info.currencyName : '---'}`
                 }
               />
-              {!refreshSure && (
-                <TouchableOpacity onPress={() => setRefreshSure(true)}>
-                  <View
-                    style={{
-                      display: 'flex',
-                      flexDirection: 'row',
-                      alignItems: 'center',
-                      justifyContent: 'center',
-                      backgroundColor: colors.card,
-                      borderRadius: 10,
-                      margin: 0,
-                      padding: 5,
-                      marginLeft: 0,
-                      minWidth: 48,
-                      minHeight: 48,
-                    }}>
-                    <FontAwesomeIcon icon={faRefresh} size={20} color={colors.primary} />
-                    {refreshMinutes > 0 && (
-                      <FadeText style={{ paddingLeft: 5 }}>
-                        {refreshMinutes.toString() + translate('transactions.minago')}
-                      </FadeText>
-                    )}
-                  </View>
-                </TouchableOpacity>
-              )}
-              {refreshSure && (
-                <TouchableOpacity
-                  onPress={async () => {
-                    setZecPrice(await RPC.rpc_getZecPrice(), Date.now());
-                    setRefreshSure(false);
-                    setRefreshMinutes(0);
-                  }}>
-                  <View
-                    style={{
-                      display: 'flex',
-                      flexDirection: 'row',
-                      alignItems: 'center',
-                      justifyContent: 'center',
-                      backgroundColor: 'red',
-                      borderRadius: 10,
-                      margin: 0,
-                      padding: 5,
-                      marginLeft: 5,
-                      minWidth: 48,
-                      minHeight: 48,
-                      borderColor: colors.primary,
-                      borderWidth: 1,
-                    }}>
-                    <FontAwesomeIcon icon={faRefresh} size={20} color={colors.primary} style={{ marginRight: 5 }} />
-                    <RegText color={colors.primary}>{translate('transactions.sure')}</RegText>
-                  </View>
-                </TouchableOpacity>
-              )}
+              <PriceFetcher setZecPrice={setZecPrice} />
             </View>
           )}
         </View>
