+++ components/Receive/Receive.tsx
@@ -1,11 +1,11 @@
 /* eslint-disable react-native/no-inline-styles */
-import React, { useContext, useState, ReactNode, useEffect } from 'react';
+import React, { useContext, useState, ReactNode } from 'react';
 import { View, Image, Modal, TouchableOpacity } from 'react-native';
 import { TabView, TabBar, SceneRendererProps, Route, NavigationState } from 'react-native-tab-view';
 import Toast from 'react-native-simple-toast';
 import { useTheme } from '@react-navigation/native';
 import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome';
-import { faBars, faEllipsisV, faInfo, faRefresh } from '@fortawesome/free-solid-svg-icons';
+import { faBars, faEllipsisV, faInfo } from '@fortawesome/free-solid-svg-icons';
 import OptionsMenu from 'react-native-option-menu';
 
 import FadeText from '../Components/FadeText';
@@ -19,7 +19,7 @@ import ImportKey from '../ImportKey';
 import SingleAddress from './components/SingleAddress';
 import { ThemeType } from '../../app/types';
 import { ContextLoaded } from '../../app/context';
-import moment from 'moment';
+import PriceFetcher from '../Components/PriceFetcher';
 
 type ReceiveProps = {
   fetchTotalBalance: () => void;
@@ -46,8 +46,6 @@ const Receive: React.FunctionComponent<ReceiveProps> = ({
   const { colors } = useTheme() as unknown as ThemeType;
   const [index, setIndex] = useState(0);
   const [routes] = useState([{ key: 'uaddr', title: translate('receive.u-title') }]);
-  const [refreshSure, setRefreshSure] = useState(false);
-  const [refreshMinutes, setRefreshMinutes] = useState(0);
 
   const [displayAddress, setDisplayAddress] = useState(uaAddress);
   const [oindex, setOIndex] = useState(0);
@@ -203,21 +201,6 @@ const Receive: React.FunctionComponent<ReceiveProps> = ({
 
   const syncStatusDisplayLine = syncingStatus.inProgress ? `(${syncingStatus.blocks})` : '';
 
-  useEffect(() => {
-    const fn = () => {
-      if (zecPrice.date > 0) {
-        const date1 = moment();
-        const date2 = moment(zecPrice.date);
-        setRefreshMinutes(date1.diff(date2, 'minutes'));
-      }
-    };
-
-    fn();
-    const inter = setInterval(fn, 5000);
-
-    return () => clearInterval(inter);
-  }, [zecPrice.date]);
-
   const renderScene: (
     props: SceneRendererProps & {
       route: Route;
@@ -348,59 +331,7 @@ const Receive: React.FunctionComponent<ReceiveProps> = ({
                   amtZec={totalBalance.total}
                   currency={currency}
                 />
-                {!refreshSure && (
-                  <TouchableOpacity onPress={() => setRefreshSure(true)}>
-                    <View
-                      style={{
-                        display: 'flex',
-                        flexDirection: 'row',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        backgroundColor: colors.card,
-                        borderRadius: 10,
-                        margin: 0,
-                        padding: 5,
-                        marginLeft: 0,
-                        minWidth: 48,
-                        minHeight: 48,
-                      }}>
-                      <FontAwesomeIcon icon={faRefresh} size={20} color={colors.primary} />
-                      {refreshMinutes > 0 && (
-                        <FadeText style={{ paddingLeft: 5 }}>
-                          {refreshMinutes.toString() + translate('transactions.minago')}
-                        </FadeText>
-                      )}
-                    </View>
-                  </TouchableOpacity>
-                )}
-                {refreshSure && (
-                  <TouchableOpacity
-                    onPress={async () => {
-                      setZecPrice(await RPC.rpc_getZecPrice(), Date.now());
-                      setRefreshSure(false);
-                      setRefreshMinutes(0);
-                    }}>
-                    <View
-                      style={{
-                        display: 'flex',
-                        flexDirection: 'row',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        backgroundColor: 'red',
-                        borderRadius: 10,
-                        margin: 0,
-                        padding: 5,
-                        marginLeft: 5,
-                        minWidth: 48,
-                        minHeight: 48,
-                        borderColor: colors.primary,
-                        borderWidth: 1,
-                      }}>
-                      <FontAwesomeIcon icon={faRefresh} size={20} color={colors.primary} style={{ marginRight: 5 }} />
-                      <RegText color={colors.primary}>{translate('transactions.sure')}</RegText>
-                    </View>
-                  </TouchableOpacity>
-                )}
+                <PriceFetcher setZecPrice={setZecPrice} />
               </View>
             )}
 
@@ -604,59 +535,7 @@ const Receive: React.FunctionComponent<ReceiveProps> = ({
                     amtZec={totalBalance.total}
                     currency={currency}
                   />
-                  {!refreshSure && (
-                    <TouchableOpacity onPress={() => setRefreshSure(true)}>
-                      <View
-                        style={{
-                          display: 'flex',
-                          flexDirection: 'row',
-                          alignItems: 'center',
-                          justifyContent: 'center',
-                          backgroundColor: colors.card,
-                          borderRadius: 10,
-                          margin: 0,
-                          padding: 5,
-                          marginLeft: 0,
-                          minWidth: 48,
-                          minHeight: 48,
-                        }}>
-                        <FontAwesomeIcon icon={faRefresh} size={20} color={colors.primary} />
-                        {refreshMinutes > 0 && (
-                          <FadeText style={{ paddingLeft: 5 }}>
-                            {refreshMinutes.toString() + translate('transactions.minago')}
-                          </FadeText>
-                        )}
-                      </View>
-                    </TouchableOpacity>
-                  )}
-                  {refreshSure && (
-                    <TouchableOpacity
-                      onPress={async () => {
-                        setZecPrice(await RPC.rpc_getZecPrice(), Date.now());
-                        setRefreshSure(false);
-                        setRefreshMinutes(0);
-                      }}>
-                      <View
-                        style={{
-                          display: 'flex',
-                          flexDirection: 'row',
-                          alignItems: 'center',
-                          justifyContent: 'center',
-                          backgroundColor: 'red',
-                          borderRadius: 10,
-                          margin: 0,
-                          padding: 5,
-                          marginLeft: 5,
-                          minWidth: 48,
-                          minHeight: 48,
-                          borderColor: colors.primary,
-                          borderWidth: 1,
-                        }}>
-                        <FontAwesomeIcon icon={faRefresh} size={20} color={colors.primary} style={{ marginRight: 5 }} />
-                        <RegText color={colors.primary}>{translate('transactions.sure')}</RegText>
-                      </View>
-                    </TouchableOpacity>
-                  )}
+                  <PriceFetcher setZecPrice={setZecPrice} />
                 </View>
               )}
             </View>
